Refactor the class to a functional component that uses hooks
